vocab-importer/requirements.txt
streamlit>=1.32.0
pytest>=8.0.0
pytest-cov>=4.1.0

# LLM APIs
groq>=0.4.2        # Groq API
openai>=1.12.0     # ChatGPT API
google-generativeai>=0.3.2  # Gemini API

# Utilities
python-dotenv>=1.0.0  # For environment variables
requests>=2.31.0    # For API calls
jsonschema>=4.21.1  # For JSON validation 